Split StaticQueryExecuteTrait from Query
